refactor(entity): rename jpaEntity to entity
