+++ src/main/kotlin/ecommerce/entity/User.kt
@@ -1,4 +1,4 @@
-package ecommerce.jpaEntity
+package ecommerce.entity
 
 import ecommerce.enums.UserRole
 import jakarta.persistence.CascadeType
