+++ src/main/kotlin/ecommerce/repository/CartStatisticsRepository.kt
@@ -2,7 +2,7 @@ package ecommerce.repository
 
 import ecommerce.dto.cartStatistics.MembersWhoAddedToCartDTO
 import ecommerce.dto.cartStatistics.TopAddedProductsDTO
-import ecommerce.jpaEntity.CartStatistics
+import ecommerce.entity.CartStatistics
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
