+++ src/main/kotlin/ecommerce/service/CartService.kt
@@ -3,9 +3,9 @@ package ecommerce.service
 import ecommerce.dto.cartProduct.CartProductResponseDTO
 import ecommerce.enums.CartAction
 import ecommerce.exception.EntityNotFoundException
-import ecommerce.jpaEntity.CartStatistics
-import ecommerce.jpaEntity.Product
-import ecommerce.jpaEntity.User
+import ecommerce.entity.CartStatistics
+import ecommerce.entity.Product
+import ecommerce.entity.User
 import ecommerce.repository.CartProductRepository
 import ecommerce.repository.CartRepository
 import ecommerce.repository.CartStatisticsRepository
